feat: Add Integration test for Repository Layer. Ref #6
