+++ src/test/java/com/roshlab/savings/AbstractSavingsApplicationTests.java
@@ -1,0 +1,25 @@
+package com.roshlab.savings;
+
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+public abstract class AbstractSavingsApplicationTests {
+
+    private static final PostgreSQLContainer<?> POSTGRES_CONTAINER;
+
+    static {
+        POSTGRES_CONTAINER = new PostgreSQLContainer<>("postgres:15")
+                .withDatabaseName("savings")
+                .withUsername("postgres")
+                .withPassword("postgres");
+        POSTGRES_CONTAINER.start();
+    }
+
+    @DynamicPropertySource
+    static void properties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", POSTGRES_CONTAINER::getJdbcUrl);
+        registry.add("spring.datasource.username", POSTGRES_CONTAINER::getUsername);
+        registry.add("spring.datasource.password", POSTGRES_CONTAINER::getPassword);
+    }
+}
