+++ src/test/java/com/roshlab/savings/SavingsApplicationTests.java
@@ -2,9 +2,11 @@ package com.roshlab.savings;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest()
-class SavingsApplicationTests {
+@SpringBootTest
+@ActiveProfiles("test")
+class SavingsApplicationTests extends AbstractSavingsApplicationTests {
 
     @Test
     void contextLoads() {
