+++ src/test/java/com/roshlab/savings/repository/SavingsAccountRepositoryTest.java
@@ -1,37 +1,21 @@
 package com.roshlab.savings.repository;
 
+import com.roshlab.savings.AbstractSavingsApplicationTests;
 import com.roshlab.savings.entity.SavingsAccount;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@Testcontainers
 @SpringBootTest
-class SavingsAccountRepositoryTest {
-
-    // Start a disposable Postgres container for tests with Testcontainers
-    @Container
-    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:15")
-            .withDatabaseName("savings")
-            .withUsername("postgres")
-            .withPassword("postgres");
-
-    // Bind Spring Boot datasource properties to Testcontainers variables
-    @DynamicPropertySource
-    static void properties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", postgres::getJdbcUrl);
-        registry.add("spring.datasource.username", postgres::getUsername);
-        registry.add("spring.datasource.password", postgres::getPassword);
-    }
+@ActiveProfiles("test")
+@Transactional
+class SavingsAccountRepositoryTest extends AbstractSavingsApplicationTests {
 
     @Autowired
     private SavingsAccountRepository repository;
